# docs/swagger.yaml
definitions:
...
  v1.ContactHistoryListRequest:
    properties:
      biz_type:
        type: integer
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
...
  v1.ContactVoucherBuyRequest:
    properties:
      sku_code:
        example: contact_voucher_5
        type: string
    required:
    - sku_code
    type: object
  v1.ContactVoucherCallbackCostRequest:
    properties:
      purpose_id:
        type: integer
      purpose_type:
        type: integer
      purpose_user_id:
        type: integer
      purpose_user_name:
        type: string
      purpose_user_phone:
        type: string
    type: object
  v1.ContactVoucherCostRequest:
    properties:
      purpose_id:
        type: integer
      purpose_type:
        type: integer
      purpose_user_id:
        type: integer
      purpose_user_name:
        type: string
      purpose_user_phone:
        type: string
    type: object
  v1.ContactVoucherRecordType:
    enum:
    - buy
    - cost
    type: string
    x-enum-varnames:
    - ContactVoucherRecordBuy
    - ContactVoucherRecordCost
  v1.ContactVoucherRecordsItem:
    properties:
      change_num:
        type: integer
      create_at:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/v1.ContactVoucherRecordType'
    type: object
  v1.ContactVoucherRecordsResponseData:
    properties:
      contact_voucher_num:
        type: integer
      list:
        items:
          $ref: '#/definitions/v1.ContactVoucherRecordsItem'
        type: array
      list_total:
        type: integer
    type: object
...
  v1.OrderStatusRequest:
    properties:
      order_no:
        type: string
    required:
    - order_no
    type: object
  v1.OrderStatusResponseData:
    properties:
      order_no:
        type: string
      status:
        description: '订单状态: 1=待支付 2=已支付 3=已取消 4=已退款'
        enum:
        - 1
        - 2
        - 3
        - 4
        example: 2
        type: integer
    type: object
  v1.PayOrderResponseData:
    properties:
      amount:
        type: number
      order_id:
        type: integer
      order_no:
        type: string
      pay_params:
        $ref: '#/definitions/v1.PayParams'
    type: object
  v1.PayParams:
    properties:
      nonceStr:
        type: string
      package:
        type: string
      paySign:
        type: string
      signType:
        type: string
      timeStamp:
        type: string
    type: object
...
  v1.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
...
  v1.UserOrderListItem:
    properties:
      amount:
        type: number
      create_at:
        type: string
      order_id:
        type: integer
      order_no:
        type: string
      paid_at:
        type: string
      product_type:
        description: '商品类型: 1=置顶 2=联系券 3=刷新 4=招租发布'
        enum:
        - 1
        - 2
        - 3
        - 4
        example: 1
        type: integer
      title:
        type: string
    type: object
  v1.UserOrderListRequest:
    properties:
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
  v1.UserOrderListResponseData:
    properties:
      list:
        items:
          $ref: '#/definitions/v1.UserOrderListItem'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
...
  /contact_voucher/buy:
    post:
      consumes:
      - application/json
      description: 传入套餐查询接口返回的 sku_code。价格、联系券数量和赠送数量均由服务端读取，支付成功后自动到账。
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ContactVoucherBuyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PayOrderResponseData'
      security:
      - Bearer: []
      summary: 联系券充值
      tags:
      - 联系券模块
  /contact_voucher/callback_cost:
    post:
      consumes:
      - application/json
      description: 消耗1张联系券用于回拨电话，不写入联系记录表，仅记录到回拨记录表。券余额不足时返回错误码 ErrInsufficientVoucher。
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ContactVoucherCallbackCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 联系券消费（回拨电话）
      tags:
      - 联系券模块
  /contact_voucher/cost:
    post:
      consumes:
      - application/json
      description: 消耗1张联系券，并可选记录联系历史。purpose_id/purpose_type/purpose_user_id/purpose_user_name/purpose_user_phone
        均为可选，传入时自动写入联系记录。券余额不足时返回错误码 ErrInsufficientVoucher。
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ContactVoucherCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 联系券消费（拨打电话）
      tags:
      - 联系券模块
  /contact_voucher/records:
    post:
      consumes:
      - application/json
      description: '返回当前用户的联系券余额和流水记录。响应中 type: 1=消费 2=充值；change_num 为变更张数（消费为负数，充值为正数）。'
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ContactHistoryListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ContactVoucherRecordsResponseData'
      security:
      - Bearer: []
      summary: 我的券包（联系券流水）
      tags:
      - 联系券模块
...
  /order/status:
    post:
      consumes:
      - application/json
      description: '前端唤起微信支付后，通过此接口确认订单在后端的真实支付状态。status: 1=待支付 2=已支付 3=已取消 4=已退款。建议支付完成后轮询直到
        status=2 再刷新业务页面。'
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.OrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.OrderStatusResponseData'
      security:
      - Bearer: []
      summary: 查询订单支付状态
      tags:
      - 订单
...
  /user/orders:
    post:
      consumes:
      - application/json
      description: '返回当前用户已支付的订单列表，按创建时间倒序。product_type: 1=岗位置顶 2=联系券 3=付费刷新 4=招租发布。amount
        单位：元。'
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UserOrderListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserOrderListResponseData'
      security:
      - Bearer: []
      summary: 消费记录
      tags:
      - 个人中心
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"